+++ modules/zuora/test/zuoraClientIntegration.test.ts
@@ -6,7 +6,7 @@
 
 import { Logger } from '@modules/zuora/logger';
 import { BearerTokenProvider } from '../src/auth/bearerTokenProvider';
-import { getOAuthClientCredentials } from '../src/oAuthCredentials';
+import { getOAuthClientCredentials } from '../src/auth/oAuthCredentials';
 import { ZuoraClient } from '../src/zuoraClient';
 import type { ZuoraSubscription } from '../src/zuoraSchemas';
 import { zuoraSubscriptionResponseSchema } from '../src/zuoraSchemas';
