+++ modules/zuora/test/zuoraIntegration.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { BearerTokenProvider } from '../src/auth/bearerTokenProvider';
-import { getOAuthClientCredentials } from '../src/oAuthCredentials';
+import { getOAuthClientCredentials } from '../src/auth/oAuthCredentials';
 import { getSubscription } from '../src/subscription';
 import { ZuoraClient } from '../src/zuoraClient';
